+++ src/rpc/rpc_util.rs
@@ -41,11 +41,12 @@ pub fn is_streaming_method(method_name: &str) -> bool {
     STREAMING_METHODS.contains(&method_name)
 }
 
-const V1_METHODS: [&str; 5] = [
+const V1_METHODS: [&str; 6] = [
     ETH_ACCOUNTS,
     ETH_BLOCK_NUMBER,
     ETH_CHAIN_ID,
     ETH_GAS_PRICE,
+    ETH_GET_BALANCE,
     STATE_MARKET_STORAGE_DEAL,
 ];